test: add e2e tests for `setupExitSignals` option (#4130)
